+++ src/main/java/es/uji/ei1027/SkillSharing/controller/RequestController.java
@@ -31,8 +31,6 @@ public class RequestController {
     private UserValidator validator = new UserValidator();
 
 
-    private UserValidator validator = new UserValidator();
-
     @Autowired
     public void setCollaborationDao(CollaborationDao collaborationDao){
         this.collaborationDao = collaborationDao;
@@ -52,11 +50,9 @@ public class RequestController {
     @RequestMapping(value="/accept/{idRequest}", method=RequestMethod.GET)
     public String accept(HttpSession session, Model model, @PathVariable Integer idRequest) {
         Student student= (Student) session.getAttribute("student");
-        if ( student == null)
-        {
-            session.setAttribute("nextUrl","/request/accept/"+idRequest);
-            model.addAttribute("student", new Student());
-            return "login";
+        String mensaje = validator.comprobar_conexion(session, model, "/accept/{id}");
+        if (!mensaje.equals("")){
+            return mensaje;
         }
         Request request = requestDao.getRequest(idRequest);
         if (student.getIdStudent()!=request.getIdStudent()){
@@ -73,14 +69,22 @@ public class RequestController {
         return "redirect:../list";
     }
     @RequestMapping(value = "/delete/{idRequest}")
-    public String processDeleteRequest(@PathVariable Integer idRequest){
+    public String processDeleteRequest(HttpSession session, Model model, @PathVariable Integer idRequest){
+        String mensaje = validator.comprobar_conexion(session, model, "/accept/{id}");
+        if (!mensaje.equals("")){
+            return mensaje;
+        }
         requestDao.deleteRequest(idRequest);
         return "redirect:../mis_demandas";
     }
 
 
     @RequestMapping("/list")
-    public String listRequest(Model model){
+    public String listRequest(HttpSession session, Model model){
+        String mensaje = validator.comprobar_conexion(session, model, "/accept/{id}");
+        if (!mensaje.equals("")){
+            return mensaje;
+        }
         model.addAttribute("requests", requestDao.getValidRequests());
         model.addAttribute("skillTypes", skillTypeDao.getSkillTypes());
         return "request/list";
@@ -88,7 +92,10 @@ public class RequestController {
 
     @RequestMapping(value="/add")
     public String addRequest(HttpSession session,Model model){
-        String mensaje = validator.comprobar_conexion(session, model, "/request/add");
+        String mensaje = validator.comprobar_conexion(session, model, "/accept/{id}");
+        if (!mensaje.equals("")){
+            return mensaje;
+        }
         model.addAttribute("request", new Request());
         model.addAttribute("skillTypes", skillTypeDao.getSkillTypes());
         return "request/add";
@@ -96,11 +103,9 @@ public class RequestController {
     @RequestMapping(value="/mis_demandas")
     public String mis_requests(HttpSession session,Model model){
         Student student= (Student) session.getAttribute("student");
-        if ( student == null)
-        {
-            session.setAttribute("nextUrl","/request/mis_demandas/");
-            model.addAttribute("student", new Student());
-            return "login";
+        String mensaje = validator.comprobar_conexion(session, model, "/accept/{id}");
+        if (!mensaje.equals("")){
+            return mensaje;
         }
         model.addAttribute("skillTypes", skillTypeDao.getSkillTypes());
         model.addAttribute("requests",requestDao.getRequests(student.getIdStudent()));
@@ -121,7 +126,11 @@ public class RequestController {
     }
 
     @RequestMapping(value="/update/{idRequest}", method=RequestMethod.GET)
-    public String editRequest(Model model, @PathVariable Integer idRequest){
+    public String editRequest(HttpSession session, Model model, @PathVariable Integer idRequest){
+        String mensaje = validator.comprobar_conexion(session, model, "/accept/{id}");
+        if (!mensaje.equals("")){
+            return mensaje;
+        }
         model.addAttribute("request", requestDao.getRequest(idRequest));
         model.addAttribute("skillTypes", skillTypeDao.getSkillTypes());
         return "request/update";
